IGNITE-28819 [4/N] Thin client implementation extraction
