+++ modules/core/src/test/java/org/apache/ignite/spi/communication/tcp/GridTcpCommunicationSpiSkipWaitHandshakeOnClientTest.java
@@ -30,7 +30,7 @@ import org.apache.ignite.testframework.junits.common.GridCommonAbstractTest;
 import org.junit.Assert;
 import org.junit.Test;
 
-import static org.apache.ignite.spi.communication.tcp.TcpCommunicationSpi.makeMessageType;
+import static org.apache.ignite.internal.util.CommonUtils.makeMessageType;
 
 /**
  * This test check that client sends only Node ID message type on connect.
